+++ controllers/helpers/parse-ingredients.js
@@ -1,0 +1,24 @@
+const { parse } = require('recipe-ingredient-parser');
+const sortIngredients = require('./sort-ingredients.js');
+
+module.exports = (cartRecipes) => {
+  const listOfIngredients = [];
+  cartRecipes.forEach((recipe) => {
+    // iterate through recipes in cart
+    recipe.ingredientLines.forEach((ingredient) => {
+      // remove symbols
+      const ingrWithoutStars = ingredient.replace('*', '');
+      const ingrWithoutPeriod = ingrWithoutStars.replace('.', '');
+
+      // parse
+      const parsedIngredient = parse(ingrWithoutPeriod);
+
+      // add to array
+      listOfIngredients.push(parsedIngredient);
+    });
+  });
+
+  sortIngredients(listOfIngredients);
+  console.log(listOfIngredients);
+  return listOfIngredients;
+};
