+++ controllers/helpers/sort-ingredients.js
@@ -1,0 +1,4 @@
+/* eslint-disable max-len */
+module.exports = (listOfParsedIngredients) => {
+  listOfParsedIngredients.sort((a, b) => ((a.ingredient.toLowerCase() > b.ingredient.toLowerCase()) ? 1 : -1));
+};
