+++ views/cart.hbs
@@ -1,6 +1,6 @@
 <div class="cart-container">
 
-	<a href='/dashboard/cart'>Cart</a>
+	<a href='/cart'>Cart</a>
 	<div class="cart-header">
 		<h1>My Cart</h1>
 		<a href="/cart/grocery-list"><button class="btn btn-success checkout-btn">Create Grocery List</button></a>
