+++ views/login.hbs
@@ -6,13 +6,27 @@
     <!--Username-->
     <div class="form-group">
       <label for="username"></label>
-      <input class="form-control" type="text" name="username" placeholder="Username" required>
+      <input 
+        class="form-control" 
+        type="text" 
+        name="username" 
+        placeholder="Username" 
+        autocomplete="email" 
+        required
+      >
     </div>
 
     <!--Password-->
     <div class="form-group">
       <label for="password"></label>
-      <input class="form-control" type="password" name="password" placeholder="Password" required>
+      <input 
+        class="form-control" 
+        type="password" 
+        name="password" 
+        placeholder="Password" 
+        autocomplete="current-password" 
+        required
+      >
     </div>
 
     <!--Buttons-->
